slight style changes
