+++ client/src/components/DropMenu.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import '../App.css';
+
+const DropDown = () => (
+  <li class="dropdown">
+    <a href="javascript:void(0)" class="dropbtn">Dropdown</a>
+    <div class="dropdown-content">
+      <a href="#">Link 1</a>
+      <a href="#">Link 2</a>
+      <a href="#">Link 3</a>
+    </div>
+  </li>
+)
+
+export default DropDown;
+
+
+
+
